Switch layout store back to PageData
This is such a mess; a store works in dev mode even with javascript
disabled (i.e., ssr is happening and generating the correct page), but
when prerendering via adapter-static we get the pre-store-update values
rendered.

Using PageData means each page that wants to modify its layout (like, to
set a title) must have a +page.ts.