Store photos in the cloud vs git-lfs
git-lfs has been a persistent hassle with Netlify, and also jj doesn't
support it. We're now storing photos in imagekit.io, and exposing them
as a (new in Astro 5) content collection via a custom loader. This still
isn't ideal, as when loading content we have to fetch all the photos
from imagekit in order to compute the best crop, but maybe we can rely
more heavily on imagekit's transformations in the future.