Use path component for photo image id
Since we now have custom pages for each image to support individual
og:images, we want to present the correct URL when zooming a given
image. Leave tags in the hash fragment, but track the selected image
id in the path component. We still handle a hash #id:... for supporting
the redirect from the individual photo page.

This is starting to get a bit overwrought; maybe I should just use view
transitions and actually have persistent detail pages.